nouvelle_ajout

+++ src/app/departements/departements.component.ts
@@ -11,7 +11,7 @@ import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router'; // Ajoutez cette ligne
 
 @Component({
-  selector: 'app-departements',
+  selector: 'app-cohortes',
   standalone: true,
   imports: [CommonModule, ReactiveFormsModule, FormsModule],
   templateUrl: './departements.component.html',
@@ -38,7 +38,7 @@ export class DepartementsComponent implements OnInit {
   constructor(
     private departementService: DepartementService,
     private fb: FormBuilder,
-    private router: Router // Ajoutez cette ligne
+    private router: Router
   ) {
     this.departmentForm = this.fb.group({
       nom_departement: ['', Validators.required],
@@ -58,12 +58,13 @@ export class DepartementsComponent implements OnInit {
     }
   }
 
+  // Charge la liste des départements
   loadDepartements(): void {
     console.log('loadDepartements called');
     this.departementService
       .getDepartements()
       .then((data) => {
-        console.log('Departements loaded:', data);
+        console.log('Départements chargés:', data);
         this.departements = data;
         this.filterDepartements();
       })
@@ -72,6 +73,7 @@ export class DepartementsComponent implements OnInit {
       });
   }
 
+  // Filtre les départements en fonction de la recherche
   filterDepartements(): void {
     console.log('filterDepartements called with query:', this.searchQuery);
     if (this.searchQuery.trim() === '') {
@@ -92,11 +94,21 @@ export class DepartementsComponent implements OnInit {
     }
   }
 
+  // Ouvre le modal de création
   openCreateModal(): void {
     console.log('Ouverture du modal de création');
     this.showModal = true;
   }
 
+  // Ferme le modal de création
+  closeModal(): void {
+    console.log('Fermeture du modal de création');
+    this.showModal = false;
+    this.departmentForm.reset();
+    this.errorMessage = null;
+  }
+
+  // Crée un nouveau département
   createDepartement(event: Event): void {
     event.preventDefault();
     if (this.departmentForm.valid) {
@@ -131,6 +143,7 @@ export class DepartementsComponent implements OnInit {
     }
   }
 
+  // Ouvre le modal de modification
   openEditModal(departement: any): void {
     console.log(
       'Ouverture du modal de modification pour le département',
@@ -144,13 +157,16 @@ export class DepartementsComponent implements OnInit {
     this.showEditModal = true;
   }
 
+  // Ferme le modal de modification
   closeEditModal(): void {
     console.log('Fermeture du modal de modification');
     this.showEditModal = false;
     this.departementToEdit = null;
     this.editDepartmentForm.reset();
+    this.editErrorMessage = null;
   }
 
+  // Met à jour un département
   updateDepartement(event: Event): void {
     event.preventDefault();
     if (this.editDepartmentForm.valid && this.departementToEdit) {
@@ -190,11 +206,7 @@ export class DepartementsComponent implements OnInit {
     }
   }
 
-  closeModal(): void {
-    this.showModal = false;
-    this.departmentForm.reset();
-    this.errorMessage = null;
-  }
+  // Ouvre le modal de suppression
   openDeleteModal(departement: any): void {
     console.log(
       'Ouverture du modal de suppression pour le département',
@@ -204,14 +216,38 @@ export class DepartementsComponent implements OnInit {
     this.showDeleteModal = true;
   }
 
+  // Ferme le modal de suppression
   closeDeleteModal(): void {
     console.log('Fermeture du modal de suppression');
     this.showDeleteModal = false;
     this.departementToDelete = null;
   }
 
+  // Vérifie si un département peut être supprimé
+  canDeleteDepartement(departement: any): boolean {
+    return departement && departement.users && departement.users.length === 0;
+  }
+
+  // Vérifie si tous les départements sélectionnés peuvent être supprimés
+  canDeleteSelectedDepartements(): boolean {
+    return this.selectedDepartements.every((departement) =>
+      this.canDeleteDepartement(departement)
+    );
+  }
+
+  // Vérifie si un département sélectionné contient des utilisateurs
+  hasUsersInSelectedDepartements(): boolean {
+    return this.selectedDepartements.some(
+      (departement) => departement.users && departement.users.length > 0
+    );
+  }
+
+  // Confirme la suppression d'un département
   confirmDelete(): void {
-    if (this.departementToDelete) {
+    if (
+      this.departementToDelete &&
+      this.canDeleteDepartement(this.departementToDelete)
+    ) {
       this.departementService
         .deleteDepartement(this.departementToDelete.id)
         .then((response) => {
@@ -222,35 +258,52 @@ export class DepartementsComponent implements OnInit {
         .catch((error) => {
           console.error('Erreur lors de la suppression du département', error);
         });
+    } else {
+      console.log(
+        'Ce département contient des utilisateurs, il ne peut pas être supprimé.'
+      );
     }
   }
 
+  // Confirme la suppression multiple de départements
+  confirmDeleteMultiple(): void {
+    if (this.canDeleteSelectedDepartements()) {
+      this.selectedDepartements.forEach((departement) => {
+        this.departementService
+          .deleteDepartement(departement.id)
+          .then((response) => {
+            console.log('Département supprimé avec succès', response);
+            this.loadDepartements();
+          })
+          .catch((error) => {
+            console.error(
+              'Erreur lors de la suppression du département',
+              error
+            );
+          });
+      });
+      this.selectedDepartements = [];
+      this.closeDeleteMultipleModal();
+    } else {
+      console.log(
+        'Un ou plusieurs départements sélectionnés contiennent des utilisateurs, ils ne peuvent pas être supprimés.'
+      );
+    }
+  }
+
+  // Ouvre le modal de suppression multiple
   openDeleteMultipleModal(): void {
     console.log('Ouverture du modal de suppression multiple');
     this.showDeleteMultipleModal = true;
   }
 
+  // Ferme le modal de suppression multiple
   closeDeleteMultipleModal(): void {
     console.log('Fermeture du modal de suppression multiple');
     this.showDeleteMultipleModal = false;
   }
 
-  confirmDeleteMultiple(): void {
-    this.selectedDepartements.forEach((departement) => {
-      this.departementService
-        .deleteDepartement(departement.id)
-        .then((response) => {
-          console.log('Département supprimé avec succès', response);
-          this.loadDepartements();
-        })
-        .catch((error) => {
-          console.error('Erreur lors de la suppression du département', error);
-        });
-    });
-    this.selectedDepartements = [];
-    this.closeDeleteMultipleModal();
-  }
-
+  // Ajoute ou retire un département de la sélection
   toggleSelection(departement: any): void {
     const index = this.selectedDepartements.indexOf(departement);
     if (index > -1) {
@@ -260,6 +313,7 @@ export class DepartementsComponent implements OnInit {
     }
   }
 
+  // Sélectionne ou désélectionne tous les départements
   toggleAllSelection(): void {
     const allSelected = this.departements.every(
       (departement) => departement.selected
@@ -270,6 +324,7 @@ export class DepartementsComponent implements OnInit {
     this.selectedDepartements = allSelected ? [] : this.departements;
   }
 
+  // Ouvre le modal de suppression multiple pour les départements sélectionnés
   deleteSelectedDepartements(): void {
     console.log(
       'Ouverture du modal de suppression multiple pour les départements sélectionnés',
@@ -278,6 +333,7 @@ export class DepartementsComponent implements OnInit {
     this.openDeleteMultipleModal();
   }
 
+  // Recherche des départements
   searchDepartements(event: Event): void {
     const target = event.target as HTMLInputElement;
     if (target) {
@@ -290,6 +346,7 @@ export class DepartementsComponent implements OnInit {
     }
   }
 
+  // Pagination
   get paginatedDepartements(): any[] {
     const startIndex = (this.currentPage - 1) * this.itemsPerPage;
     const endIndex = startIndex + this.itemsPerPage;
@@ -306,7 +363,7 @@ export class DepartementsComponent implements OnInit {
     }
   }
 
-  // Ajoutez cette méthode pour naviguer vers la page de détails du département
+  // Navigue vers la page de détails du département
   viewDepartementDetails(departementId: number): void {
     this.router.navigate(['/departement', departementId]);
   }
